Updated the FactoryProduction window to include Clear Selection and Idle buttons.
Factories can now be set into an idle state instead of the user having to click outside of an icon in the production grid. Also a clear selection button has been added to make clearing the selection more obvious.

Structure now sports a new forceIdle() function that forces a structure to remain in and Idle state unless the user manually reenables it. If a structure later becomes disabled/destroyed, the user will not be able to enable the structure via forceIdle(false).

git-svn-id: https://svn.outpostuniverse.org:8443/svn/outposthd@325 bdf9137d-f3be-7946-88fe-7b0bbfdbf31a