When the length() function of the slider is used to set the length. If the current position is above the new maximum value. The current position is set to max value.
When moving the slider, there is a extended area where the slider will be moving even if out of the slider control.
This area is 1 width/height on each sides of the slider.
For a slider of 10 pixels wide the global sliding area will be 30 pixels wide.


The ListBox now use pixel based movement when the slider is displayed.

git-svn-id: https://svn.outpostuniverse.org:8443/svn/outposthd@341 bdf9137d-f3be-7946-88fe-7b0bbfdbf31a