Use `auto` in template method to hold return value
The actual type is potentially unknown, since it is indirectly dependent
on the template parameter `T`. It's possible `T` could declare a
function with the same name and signature, but return a different type.

This also gets around Linux compiler errors, since the actual type
hasn't been declared yet. With the current code structure, the type is
not visible until after the include of the header containing the
template method.