Fixing a wild pointer dereference bug.
This fix feels more like a workaround than an actual fix. Reality is the interactions of everything going on is more complicated than I realized... the ListBox's appear to be responding to events even though they're hidden (probably related to changing things around in how event handling code responds in GUI controls).

Anyway, this isn't a great solution and fails to address the root cause but by forcing the UI panels to refresh their lists upon loading of a saved game from within the GameState it does avoid the issue.