Replace ternary with `if`/`else`
This is maybe not the best place to use a ternary operator. Often the
ternary operator is used to conditionally return one of two values.
Sometimes it is used to conditionally execute one of two statements.
Usually in those cases, the statements are quite short and highly
similar to each other. Here, the statements are overflowing the line,
and are kind of different from each other. It seems a little more clear
to use an `if`/`else` block instead. Kind of subjective call though.