Remove compiler flag to generate debug info
Currently the debug information is unused (and rather large). In
particular, it is not needed for CI builds, and is generally not needed
for local development.

Instead, debug information can be generated by passing the `-g` flag in
the `CXXFLAGS.EXTRA` parameter. This parameter can be set in an
environment variable if someone wishes to always generate debug
information.