Update NAS2D for updated Mingw-w64 Docker image
Technically I don't need to update the submodule to access this, since
CircleCI will find the image through DockerHub, where it is already
uploaded, independent of the submodule. However, the instructions for
building the image are found in the submodule, so it seems a bit weird
to not update the submodule before using the image.