Remove old comment
Table lookup is not likely to improve performance here. The modulus
operator isn't that slow (same as integer division), and that's only for
the general case. For a power of 2, as used here, and even the most
rudimentary optimization setting, the modulus will compile to a bit
mask, which is about the fastest operation you can have. It would be
much faster than memory access needed for a table looking, and even more
so if there was a cache miss.