Use returned result, rather than file global
As there are no longer external references to any of the variables, move
them into the function as static variables.

These two operations are done together to prevent warnings about
variable name shadowing, which would otherwise exist if the original
file level global was still in scope when the new local variable of the
same name was introduced.