Bug fix condition for CC exists and use vectorized form
A CC location of {0, 0} means no CC has been placed. Instead of checking
this condition exactly, it was checking there was no CC along the left
or top edge of the map. A somewhat trivial distinction since the CC can
never be placed along the edge of the map anyway.