Collapse nested if to nested ternary
Mark variable as `const` now that it is initialized in a single step.

Although much more compact, I'm a little hesitant to use nested ternary.
Nested ternary is usually not great for readability. There does seem to
be a high degree of symmetry here though, so perhaps that keeps it
readable.