Use data drive loop for resource displays
There was some irregularity with the spacing. This should probably be
resolved at some point. One caution is the click event handling is not
tied to the graphics display, so the hit test area must be updated along
with the graphical updates. That is a concern which is external to the
drawResourceInfo method.

There was some irregularity with the definition of `color`, which was
defined in the first block.