Replace virtual with override in derived classes
One oddity noted, was the ListBoxBase class, which overrides the update
method from the parent class with a pure virtual method. This
effectively removes the default implementation from the base class, and
forced derived classes to reimplement. I'm uncertain if this was
intentional or not. I did however noticed all derived classes do
override this method with their own implementation, so it may have been
intentional.