Move file level globals to function level statics
This delays initialization until the first time the function is run.
This delay allows for the `Color` constant values to be initialized
first, allowing copies to be made of valid values. Without this delay,
undefined static initialization order could allow the color table to be
initialized before the `Color` constant values had been initialized,
leading to a table with all 0 entries (transparent black) for the named
color constant values.

This change impacts performance slightly, as the compiler will insert a
check before initializing the static definitions, to ensure they are
only initialized the first time the function is called. With the current
structure though, it is necessary for correctness.