Change getOreCount to use int for depth
There doesn't seem to be an inherent reason for depth to be unsigned. It
has an obvious zero point at surface level, though you could potentially
move in either direction from there. I suppose it's possible the game
could some day be extended up into the clouds. Though the current design
is to only go one direction, down into the ground.

The choice as to which way to go with the type is a bit arbitrary. It
seems using a signed value is the smaller change. It also seems to allow
for more future extensibility with skyscrapers. Perhaps go with signed
for now.