Move Control::size set method to implementation file
The original method it was delegating to called onSizeChanged 3 times.
Once after setting each respective dimension, and then again after both
dimensions had been set. Rather than delegate to this method, it may be
better to reimplement, so each dimension is set, and then a single
onSizeChanged event is raised.