Remove virtual from Slider::size
The base class `size` method is not `virtual`. If anything, the base
class method should be marked `virtual`, and this one marked as
`override`. Of course, that only matters if Slider is sized through a
pointer or reference to the base class. Current calls are made on direct
object instances, so the compiler knows the exact type, and can make
direct calls, rather than through the virtual function table.