Remove std::clamp from random number generation
The distribution will control the range, so values shouldn't need to be
clamped. Instead, the distribution can be adjusted to control the range
of values being generated.

Note there is a bit of a range bug fix here, since the clamping was
slightly more restrictive than the distribution range. This update
allows mines slightly closer to the edges, particularly the bottom right
edges, though still keeps them a few tiles away from the edges.