Remove ListBox feature to sort items when added
This is a rather inefficient way to do sorting. It would be more
efficent to pre-sort the list, or wait until all items are added and
then call sort.

I recommend control over sorting the ListBox contents be handled through
external means.

Since the sort feature is not currently used, now would be a good time
to change it.