Fix CheckBox initialization order bug
The `Font` must be set before `text` is set, as the `text` setter uses
the `Font` to calculate text size and set the `Control` size.

If the `text` was empty, the `Font::width` call would exit early and
return 0, and so the `Font` member variables were never accessed,
avoiding crashes.