Remove MainMenuOptions dialog class
This class isn't being used. It's repeatedly coming up in refactoring
maintenance tasks. Rather than burdening refactoring, we should remove
it for now.

The class isn't necessarily bad or anything, just unused and kind of
getting in the way often enough, while also not being overly complex,
and so it should be easy enough to recreate later using the original as
a guide. In terms of future changes, hopefully we'll have layout
managers by then, which will reduce a large bulk of the layout code.
Additionally, the NAS2D Configuration class has been undergoing changes,
which may impact how the settings are loaded and saved.