Update serialize order to match memory storage order
A matching storage and access pattern leads to better efficiency, and
smaller memory working set size for a large array.

The save/load code is structured in such a way that the file order
doesn't matter. Each tile is stored with its {x, y} coordinate, which is
used during loading to set the correct tile. This allows tiles to be
saved and loaded in any order.