Redeclare mPlayerResources as StorableResources instead of ResourcePool
This looks like a lot of changes and sure, lots of files were changed. I was going to rebuild everything but opted to just comment out the parts that break. Most of the changes here are just taking advantage of the new StoragableResource -- the rest will be fleshed out better and cleaned up.