Add guard for empty list increment and decrement
For increment, the guard is necessary to maintain NO_SELECTION
correctness. For decrement, the guard is only necessary if NO_SELECTION
was defined to be something other than -1.

If the NO_SELECTION criteria was instead changed to list size() (just
past the end of the list), the criteria above would be reversed.