Make IconGrid::selection unset the selection when passed invalid index
This seems to make more sense than simply ignoring an invalid index.
Trying to set an invalid index should unselect any selected item.

Additionally, if the list is modified in an unsafe way, perhaps through
a future update, such that the current selection is out of bounds, this
will correct the situation, rather than leave the selection in an out of
bounds state. That might be useful if removing a number of items from
the grid, and then selecting based on previous position.