Add production resources loading
As a note, this code is vastly simplified and expects that the values in the savegame are correct. Since they're generally only written by the software itself, it can be safely assumed that the values are indeed correct. If they're not exceptions get thrown and can be caught later on for a more graceful load but that's the subject of a different changeset.