Move exception check to after switch
Removing the `default` clause means the compiler can provide better
warnings for missing cases. If a new structure type is added, the
compiler can then mark the `switch` block as needing an update.

This also catches cases where an explicit Tube entry could lead to a seg
fault due to use of a nullptr.