Fix construction order bug for productImages map
The Image objects can't be loaded until the Filesystem object is
initialized. That makes it rather impossible for a global data structure
to contain Image objects. Instead, use a method with a static local
constructed variable, which allows constructed to be delayed until first
use. At the point the Filesystem object will be initialized, allowing
the Image objects to be loaded.