Return Color objects by-value
The Color struct is meant to be a simple value type, which is cheap to
copy. The size of a Color struct is at most the size of a reference, so
there are no gains to return by reference. Using by-value is not only
the same size (on 32-bit) or smaller (on 64-bit), but also eliminates an
indirection, allowing for faster access to the data.