Add virtual destructor to Structure class
Not sure why this was omitted -- I thought I'd added it way back in the beginning. Since this class has virtual functions, is inherited from every other structure in the game -and- gets deleted through a pointer to base class, the virtual function is a requirement for proper destruction of the object.