Remove unused method remoteItem
We probably do want a method like this in the long run. Currently
though, the parameter is dependent on the ListBoxItem type, so not very
generic or portable between the ListBox classes. Since it's not being
used, it can be removed for now, and a replacement added later, after
merging the classes.