Implement operator+ in terms of operator+=
Operator += has no need to create a new temporary object, so can be more
efficient than operator +. As such, it is generally advisable to
implement operator+ in terms of operator+=, rather than the other way
around. This avoids unnecessary construction, copy assignment, and
destruction of a temporary object in some cases.

Additionally, the operator+ implementation can use pass-by-value
semantics for the left-hand-side as a way of creating a copy, which is
then used with the self modifying operator+=. This has benefits when
used on temporaries, as rvalue semantics can potentially move a
temporary into the left-hand-side, rather than creating a full copy.
That means you only actually pay for the copy when it's strictly
required, such as when applying operator+ to a named lvalue.

Note: To use pass-by-value semantics, operator+ should be a non-member,
as member functions would use pass-by-reference for the hidden `this`
pointer.