Rename `empty()` to `isEmpty()`
The name "empty" is ambiguous in that it can be either a verb or an
adjective. Normally functions and methods are named to be verb phrases,
so the verb form may be assumed, which would imply behaviour matching
"clear". Instead, "isEmpty" clearly implies the adjective form of
"empty", and is understood to be a (usually const) method that returns a
boolean value indicating if a set does not contain anything.