Remove `const` from `mPlayerResources`
On second look, that was kind of heinous. :P

The field was described as "current player resources", implying it was
mutable. I suppose it could be declared as both `const` and `mutable`,
but aside from that being a bit of an oxymoron, it would not be in the
context normally reserved for use of `mutable`. Better to just not
declare the field as `const`. After all, by virtue of it being modified
in that one method, it is not actually const. Plus it is already marked
as `private`, so it protected from modification outside the class. The
only code surface area that matters is to not accidentally modify it
from within the class.