Delegate to operator on StorableResources rather than std::array
Use operator < and operator <= on StorableResources rather than
std::array. This is probably what was intended, as changes in < should
have corresponding effects on > and likewise for <= and >=.

This also affects the ability to declare the methods as `constexpr`,
enabling compile time computation. The `std::array` method was not
marked as `constexpr`, so affects `constexpr` of any calling method.