Add default copy and move constructors to RadioButton
Because of the custom destructor, auto generation of these methods is
deprecated. The methods are not technically needed, since the code is
written such that `RadioButton` objects are never copied or moved. In
particular, the std::vector::reserve call ensures the vector is never
resized, which would require copy/move operations to be supported. At
any rate, because of the use of std::vector to store RadioButton, these
methods need to exist.