Insert storage lists into new list
This allows the two input lists to be of different types. In particular,
they can be `std::vector` of a subtype of `Structure*`, such as
`CommandCenter*` or `StorageTanks*`. Elements of such collections can be
copied into a `std::vector<Structure*>`. It would however not be
possible to copy a `CommandCenter*` into a collection of
`StorageTanks*`, or vice versa.