Remove origin from offset for WarehouseReport lists
These values are meant to be offsets from the origin, so it makes little
sense to add the origin. Plus, the origin value has not been set at
this point, so at best it defaults to zero, and at worst it's random
garbage, which is possibly subtracted out when the origin is set.