Limit variable scope to loop in MapViewState::readRobots
Rather than manually reset each variable to 0 each iteration, move the
declaration and initialization to 0 of each variable into the loop. This
has the added benefit of limiting variable scope, which is generally a
good practice.

This is also the only place where `addRobot` is called with an explicit
id.