Update NAS2D submodule for stop frame sentinel value update
Animations can now use a `frameDelay` of `0` to indicate a stop frame,
as well as the older value of `-1`. A value of `0` doesn't have a valid
meaning as a delay, so makes sense to use as the sentinel value,
particularly since the field is declared as `unsigned`.

Use of `-1` to indicate stop frames should be considered deprecated, and
support may be removed in the future.

Additionally, "delay" in XML data is now optional, with a default value
of 0.