Allow both workers and scientists to die in the same turn
Since deaths were split between workers and scientists, their death rate
was effetcively halved. That's roughtly equivalent to doubling the
divisor. The split wasn't quite even though, with a 45%/55% split, so
that divisor should be offset by some amount. A +-5% out of 50% would
translate to an approximate 10% offset from base. After doubling the
divisor, and using the constant as an approximate base, that corresponds
to roughtly a constant offset of 50.