Add mScanPath to track what folder is being presented
This adds consistency between the folder display and what folder is
opened. It also allows the folder to track the actual subfolder being
used. Previously it was opening the parent folder.

There is a built in assumption the scan folder is relative to
`prefPath()`. This of course depends on the actual mounts used.
Additionally, the search path means the files could potentially come
from multiple folders.