Convert mStructureCostTable from std::array to std::map
This table has gaps, so `std::map` may be more appropriate here.
Additionally, using `std::map` makes it easier to convert the lookup key
from `enum` to `enum class`, as it would no longer need to be explicitly
converted with a `static_cast` to a `std::size_t` for lookup.