Move call to updateScrollPosition() to TextField::update()
Technically the old position of the call could result in some lag in
scrolling the text box due to the cursor blinking delay. Previously it
would only update the scroll position when the cursor was currently
blinking on. Additionally, it would only update the sroll position when
the control had the focus, which mean external calls to change the text
contents might not update the scroll position until the control got the
focus.