Eliminate half constants
This fixes a drawing issue where the y-axis was being squished and
losing pixels over time. This was due to a rounding issue, where the
half height values were pre-rounded and then used as a multiplier to
calculate tile position. Instead, the multiplication should be done
first, and the division and rounding can be done at the end to prevent
some of that squishing.