Remove useless bounding box check in updateTileHighlight
This check serves no real purpose. If the mouse is outside of the
visible bounds, then the calculated tile coordinates will also be
outside of the visible bounds. This is not a problem as all other code,
such as mouse highlight drawing, and click handlers will always check
for a valid visible position anyway.