Use heuristic for large increment
Often a scroll bar is used for pixel level scrolling of some container
it is attached to. In such cases, the scroll bar height is equal to the
view area height. For pixel level scrolling, the total container size is
generally the max value plus the view size (assuming there is data to
scroll). In effect, when scrolled to the max position, the end of the
container matches up with the end of the view.

The large scroll increment should roughly match up with one view area
height. Assuming pixel level scrolling, this is the view size, which
matches the scroll bar height.

Note for future: In some cases, a bit of scroll context is desired, so
it may make sense to scroll by a bit less than the view size. This may
depend a bit on what is displayed though. Roughly though, about the
height of one line of text would be a reasonable scroll context overlap
size. This should probably be a minimum of at least one line of text.
That of course assumes scroll context is desired, which maybe it's not.