Add extra `std::clamp` check to `MapView::viewSize` setter method
Potentially some day there may be some really high resolution monitors,
which might be able to view the entire map all at once. In that case,
clamp the view size at the smallest map dimension. That means the view
diamond, which is a square section of tiles, can still fit entirely
inside the map, and prevent accessing outside of the map bounds when
drawing the map.

Add more commenting about why the checks are being done.