Explicitly ignore return value from `std::system` call
According to the documentation, the return value is implementation
defined, so it's pretty difficult to use it in any sort of meaningful
way. For that reason, we will explicitly ignore it here.

Documentation:
https://en.cppreference.com/w/cpp/utility/program/system

Considering how this function is used by the two callers, this means
potentially silently ignoring errors during click event handling for the
main menu "Help" button and the "Open Folder" button from the saved game
dialog.