Make `RadioButton` derive directly from `Control`
The `text` field was already being delegated to the contained `Label`,
so it wasn't actually using the field from `TextControl`.

Additionally, there were `text` methods in the derived class that would
hide the original methods. On a related note, the inherited `text`
methods were not `virtual`.

Effectively, it was a lie to say the `RadioButton` was derived from
`TextControl`, since those derived methods would not have functioned
correctly.