Use depth from `centerTile` parameter
All calls to this overload had an assumed depth of 0, since they dealt
with surface structures. But in that case, the passed `centerTile`
parameter would be a tile with depth 0. It makes sense to just use the
tile depth here, and then this overload would work for underground
structures too.