Add `matrix.platform` to the cache `key`
This means `vcpkg` dependencies for `x86` and `x64` will be cached
separately. This is important, since we'll now only be installing a
single set of dependencies when doing builds for each `PLATFORM`
setting. Hopefully this leads to faster cache restore (less data for a
single platform), and faster cache rebuilds (only dependencies matching
the current `PLATFORM` need to be installed).