Use `NAS2D::mapToVector` for single step initialization
Here the lambda is used to deduce the correct function overload. There
is both a `std::string` and `std::wstring` overload, and there is not
enough context when deducing template parameters to `mapToVector` to
determine which one is desired.

Perhaps if `mapToVector` was updated with C++20 Concepts, it might be
possible to make this work without the lambda.