Remove explicit `default` constructor for `ResearchTracker`
With compile flag `-Wdeprecated-copy-with-dtor`, newer versions of Clang
give an error when an explicit `default` is used. The problem is the
`ResearchTracker` object is returned by-value from a function, which
requires an implicit copy constructor, and specifying an explicit
`default` constructor prevents implicit creation of the copy constructor
needed to return the object by-value.

An alternative might be to also be explicit about the copy constructor.