Be explicit about `Point` objects when calling `contains`
It would be helpful to have an overload of `contains` that can check for
`Rectangle` containment in addition to `Point` containment. However,
both `Point` and `Rectangle` are constructable from 2 values (with
`Rectangle` using a default of `0` for unspecified width and height).
This creates ambiguity over which overload to call, which results in a
compiler error. By being explicit about what type of object is being
passed to `contains`, we can add an overload without creating such
ambiguity.