Preconstruct `DelegateX` before calling `connect`/`disconnect`
This means fewer overloaded methods will be needed on the `SignalSource`
object. It also more tightly associates `this` pointers with member
function pointers by composing them into an obvious compound object.

One reason to want to reduce `connect` overloads, is to add debugging
capabilities to the `connect` method. This is easier when only a single
overload needs to be updated.