Add global functions to set and get `meanSolarDistance`
This value really has nothing to do with the `StructureCataglogue`, so shouldn't be stored there. For now, a global function seems the most convenient. Long term, perhaps the `MapViewState` object can be broken up, and one of the sub-objects can be responsible for this. Though we might still need global access to the object in that case.