Allow progress bars with zero progress to render
This is of course quite safe to do, and draws the expected empty progress bar. This does make the method slightly less safe to call in that passing a value of 0 for both `value` and `max` will result in an exception, rather than not drawing anything. Ideally though, a `max` value of 0 should never be used, and should indicate an error at the call site.