Fix sign conversion warning in `NotificationArea`
This one is a little gross, since iterator arithmatic requires a `std::ptrdiff_t` type, which is signed. There is no easy way to convert a `std::size_t` index value into an interator without a `static_cast`. Seems like there should be a better way for a `std::vector` to convert an index value into an iterator.