Add helper `hasIdleRobot`
This will be important for const correctness later, as we don't need write access to a list to determine if a robot is available. This will matter if the lists are changed to own the robots they contain, as the const-ness of the list will impact the const-ness of the returned objects.