Merge and simplify code for `getIdleRobot`
We will of course want write access to the returned robot. As such, we should not require `const` on the list, as it will force `const` and the object it returns. This isn't a problem when the list holds pointers, but will be an issue when the pointers are converted to full objects.