Add pre-check to `insertRobotIntoTable` (with caveat)
We can't really check while loading saved games, since robots are loaded before structures, as they need to be for the ID-to-robot mapping to work. That means during load, robots will be placed, before any supporting structures have been created.

To work around the load situation, we only perform the check when the max control is greater than 0. Basically, we expect a command center to be up, and to stay up by the time robots are available.