Use new `readResourcesOptional`
Remove old helper function `loadResorucesFromXmlElement`.

Technically, the old helper function would leave the out parameter unchanged when there was no sub-element. The new code implicitly zeros out the result when there is no sub-element. Considering this is used just after a new `Structure` object is initialized, this is a distinction without a difference.