Refactor `rebuild` mechanism
The `Structure` should be responsible for the rebuild mechanism, including the number of turns it will take. This should not be an external parameter.

Currently `rebuild` only applies to Roads, which normally have a construction cost of 2 turns, reduced to 1 turn in the case of a rebuild.

To match the behaviour of needing fewer turns to rebuild a road, and avoid having to modify the `turnsToBuild` property which will become read-only, we can instead modify the `age` so it is reset to `1` instead of `0`. This means the structure will complete the rebuild more quickly, though also has the implication that it will reach `maxAge` a turn early.