Revert "Temporary changes to dump current Structure values"
This reverts commit aa14b5ef76e8952e959777ea72485cf81777afe0.

We don't need those changes permanently in the project. We just want them to be part of the history, in case we ever need to modify and re-run the dumping process. Going forward, `StructureType` data will be loaded from an external `StructureTypes.xml` data file, which the previous commit could produce, with current game values.