Remove `setFrame` calls
The current animation data only has a single frame for the construction animation. That makes it pointless to set the frame here, as they all end up using the same single frame anyway.

In the future, if we want an animation to start in a random position, so not all sprites are synchronized, then we should make that part of the sprite data.