Allow `MiningFacility` to be connected to Command Centers
The `MiningFacility` doesn't need a Command Center connection, but can now transmit a connection, if it's adjacent to another `Structure`.

Technically this change isn't specific to the `MiningFacility`, but rather to any structure on the same `Tile` as a `Mine`. Normally we would only expect a `MiningFacility` on such a `Tile`.