Fix global `selfSustained` function
Before, this method could not query the `Structure` class, since it's used when placing a `Structure`, before the class is instantiated. That meant maintaining a separate function that needed to be kept in sync with the `Structure` data. No surprise, it wasn't, and so the result of this function did not necessarily match how some structures behaved.

Now that the info has been moved to `StructureType`, we can query the same info as `Structure`, even before the `Structure` class has been instantiated.