Ensure this `makefile` never tries to build NAS2D lib directly
We'll need to add a rule to build static library files to this makefile for when we split the OPHD project. However, a generic rule would pick up on the NAS2D library target and may try to build it, using the limited dependency information available in this makefile. That would lead to errors. Instead, we explicitly state there is no recipe to build this particular target.

For NAS2D, we have dependency rules that will chain to the makefile for the NAS2D project.