Extract section for default project flags
These will likely be shared by projects. In particular, we may want to define a project before the OPHD project which makes use of these flags. We would want to do that for static libraries extracted from OPHD, since OPHD will then depend on those libraries. To support that, variables for the static library output will need to be defined before OPHD project variables. That happens most naturally if the static library project is defined before the OPHD project. That means the compile flags will also need to be defined before the OPHD project, as they'll need to be defined before any static library projects.