Remove library dependencies on NAS2D
When executables are built, using these libraries, they also need to link against NAS2D, which is correctly done. However, when creating the libraries themselves, there is no need for such a dependency. In fact, listing such a dependency creates slightly wrong output, since the NAS2D library then gets packaged inside the archive for the `libOphd` and `libControls` libraries. This is incorrect, as only object files should exist in the library archives.

On newer version of MacOS, including other libraries inside library archives leads to an error about the inner library archive not being a mach-o file. (The inner library is not an object file).