Convert `mPlanets` to hold `Planet` directly
Removes uses of `new` and `delete`.

Since `Planet` doesn't support the neccesary constructor and conversion operators to support `std::vector::push_back`, we instead need to use a `std::vector` constructor to load all values. The `std::vector` constructor can use the `Planet`'s conversion constructor from `Planet::Attributes`.

Note that we need the `attributesToPlanets` helper function here, since using this constructor requires repeated use of the input collection to get both `cbegin` and `cend` iterators. The requires a named variable to hold the input collection, so we don't need to parse and load it twice. Since we eliminated the class field that previously held that value, we need a local variable to hold it.