Update copy of `vcpkg` cache from OPHD to NAS2D
The previous copy command was copying the contents of the first folder into the second folder. That effectively stripped off the root folder name from the source, losing one level of folders. That of course made the copy rather useless, which meant no time was saved when trying to build the NAS2D project.