Eliminate GitHub Actions Cache Save warning
Normally when a cache is restored with a hit of the primary key, we don't expect any changes to the output, so there should be no reason to save the cache. In fact, if we try, we'll get a warning. This shows up in both the cache save step, and in an annotation on the workflow run summary page. We don't really want to see that. Hence, we disable saving the cache when we know it would produce a warning.

For the OPHD incremental build cache, we skip the restore for the default branch. Hence we always want to save the cache when the restore step is skipped. That way we save a cache for branches to build incrementally from.