Revert "Set `VcpkgManifestInstall` to `false`"
This reverts commit 69c4f91aa7f58aba282d083159ae700b15d14991.

There were build errors after setting this. It seems it interfered with finding needed SDL headers. It wasn't noticed on the branch due to build caching, which meant nothing needed to be rebuilt, hence no failures because it couldn't find any headers.

Documentation:
https://learn.microsoft.com/en-us/vcpkg/users/buildsystems/msbuild-integration#vcpkgmanifestinstall-install-vcpkg-dependencies

> `VcpkgManifestInstall` (Install Vcpkg Dependencies)
> This property can be set to `false` to disable automatic dependency restoration during project build. Dependencies must be manually restored via the vcpkg command line separately.

Despite the description, it seems to disable other features of `vcpkg`, such as the ability to import `vcpkg` supplied header files.