Store `Difficulty` using `const&`
By making this a reference, we can no longer update what is points to, so we need to remove the `difficulty` setter method. We also don't need the setter method, since storing a reference to the original value means we always have access to the original value. We no longer need a setter method to update shadow copies of the original value.