Add forward declares for `enum class` constants used in `Common.h`
Note that we can only forward declare `enum class` and not old style `enum`. This is because no size is specified for the old style `enum` declarations. A size must be specified for a forward declaration to be possible. For `enum class` a default size of `int` is used when not explicitly specified.