Rename test file to end with `.test.cpp`
Often the test code and the code being tested are both opened together in an IDE. Normally the tabs display the filename. By making the filenames distinct, it can be easier to keep track of which tab is which.

This is also consistent with the convention used in NAS2D.