Replace `std::enable_if` with `concept`
This produces a better error message when the constraint is not satisfied.

Note that we can't create a unit test to ensure that certain code doesn't compile. Hence we can't really add a unit test for this change, at least not one that passes. Though we can temporarily add such a unit test in the next commit to document what a failure demo would look like. It can be removed after, and remain part of history without being an active failing unit test.

Switching from `std::enable_if` to `concept` changes the error message from:
> error: no type named ‘type’ in ‘struct std::enable_if<false, RandomNumberGenerator_NonNumericType_Test::TestBody()::NonNumeric>’

To:
> "constraints not satisfied" ...
> "required for the satisfaction of ‘IsNumeric<T>’" ...
> "expression ‘is_arithmetic_v<T>" ... "evaluated to ‘false’"