Use `CPPFLAGS_EXTRA` to pass pre-processor flags
The use of `WARN_EXTRA` was a holdover from an earlier version of the `makefile`, which had fewer customization variables. It was an inapprioriate variable for the flag being passed. It wasn't even for the right stage of compilation, since `WARN_EXTRA` is for compiler flags, rather than pre-processor flags.