Avoid linking SDL2 to `testLibControls`
Unit tests should avoid initializing SDL or doing any drawing.

The `sdl2-config` link flags cause the `SDL2main` library to be included, which hijacks the `main` function. The flags are setup such that the SDL flags come before the link flag to include the `gmock_main` library. This results in linking in `main` from `SDL2main`, which references `SDL_main` as the renamed `main` function, which doesn't exist, and causes a link error. Avoiding this library should avoid the problem.