Remove dummy file `libOPHD.cpp`
This file was just an empty placeholder from when the project was first created and no other source files existed. The project needed at least one source file in order to compile it. The placeholder file is no longer needed now that other source files have been moved to this project.